refactor: 당첨 결과를 계산하는 메서드를 WinningResult 로 이동
