+++ src/main/java/lotto/domain/WinningResult.java
@@ -9,8 +9,12 @@ public class WinningResult {
 
     private final Map<Rank, Integer> result;
 
-    public WinningResult(Map<Rank, Integer> result) {
+    private WinningResult(Map<Rank, Integer> result) {
         this.result = result;
+    }
+
+    public static WinningResult of(Lottos lottos, WinningLotto winningLotto) {
+        return new WinningResult(mapResult(lottos, winningLotto));
     }
 
     public BigDecimal calculateProfitRate(int buyCash) {
@@ -24,6 +28,34 @@ public class WinningResult {
             prizeMoney = prizeMoney.add(new BigDecimal(result.get(rank) * rank.getPrizeMoney()));
         }
         return prizeMoney;
+    }
+
+    private static Map<Rank, Integer> mapResult(Lottos lottos, WinningLotto winningLotto) {
+        Map<Rank, Integer> result = initResult();
+
+        for (Lotto lotto : lottos.get()) {
+            Rank rank = getRank(winningLotto, lotto);
+            if (rank == Rank.NONE) continue;
+            result.put(rank, result.get(rank) + 1);
+        }
+
+        return result;
+    }
+
+    private static Map<Rank, Integer> initResult() {
+        Map<Rank, Integer> result = new LinkedHashMap<>();
+        for (Rank rank : Rank.values()) {
+            if (rank == Rank.NONE) continue;
+            result.put(rank, 0);
+        }
+
+        return result;
+    }
+
+    private static Rank getRank(WinningLotto winningLotto, Lotto userLotto) {
+        int matchCount = winningLotto.getMatchCount(userLotto);
+        boolean matchBonus = winningLotto.getMatchBonus(userLotto);
+        return Rank.find(matchCount, matchBonus);
     }
 
     public Map<Rank, Integer> getResult() {
